Split Android prompt flow packages
