+++ apps/android/app/src/main/java/com/flashcardsopensourceapp/app/di/AppGraph.kt
@@ -5,15 +5,15 @@ import android.util.Log
 import androidx.lifecycle.ViewModelStore
 import androidx.lifecycle.ViewModelStoreOwner
 import com.flashcardsopensourceapp.app.AutoSyncController
-import com.flashcardsopensourceapp.app.FeedbackPromptController
-import com.flashcardsopensourceapp.app.GuestSignInAfterReviewPromptController
-import com.flashcardsopensourceapp.app.feedbackPromptIdentityKey
 import com.flashcardsopensourceapp.app.navigation.AppPackageInfo
 import com.flashcardsopensourceapp.app.navigation.loadPackageInfo
 import com.flashcardsopensourceapp.app.ProgressContextRefreshController
-import com.flashcardsopensourceapp.app.SharedPreferencesFeedbackPromptStore
 import com.flashcardsopensourceapp.app.observability.renderSanitizedThrowableLogFields
-import com.flashcardsopensourceapp.app.SharedPreferencesGuestSignInAfterReviewPromptStore
+import com.flashcardsopensourceapp.app.prompts.feedback.FeedbackPromptController
+import com.flashcardsopensourceapp.app.prompts.feedback.SharedPreferencesFeedbackPromptStore
+import com.flashcardsopensourceapp.app.prompts.feedback.feedbackPromptIdentityKey
+import com.flashcardsopensourceapp.app.prompts.guestreview.GuestSignInAfterReviewPromptController
+import com.flashcardsopensourceapp.app.prompts.guestreview.SharedPreferencesGuestSignInAfterReviewPromptStore
 import com.flashcardsopensourceapp.app.store.NoOpStoreReviewAnalyticsReporter
 import com.flashcardsopensourceapp.app.store.StoreReviewActivityProvider
 import com.flashcardsopensourceapp.app.store.StoreReviewRequestManager
