+++ apps/android/app/src/main/java/com/flashcardsopensourceapp/app/prompts/guestreview/GuestSignInAfterReviewPromptDialog.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.app
+package com.flashcardsopensourceapp.app.prompts.guestreview
 
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Text
@@ -7,6 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
+import com.flashcardsopensourceapp.app.R
 
 const val guestSignInAfterReviewPromptTag: String = "guest_sign_in_after_review_prompt"
 const val guestSignInAfterReviewPromptSignInButtonTag: String =
