+++ apps/android/app/src/main/java/com/flashcardsopensourceapp/app/prompts/guestreview/GuestSignInAfterReviewPromptPolicy.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.app
+package com.flashcardsopensourceapp.app.prompts.guestreview
 
 import com.flashcardsopensourceapp.data.local.model.cloud.CloudAccountState
